fixed some bugs

+++ controller/product.js
@@ -1,5 +1,6 @@
-const Product = require("../model/product");
+const ObjectID = require("mongodb").ObjectID;
 const _ = require("lodash");
+const Product = require("../model/product");
 
 class ProductController {
   async getAllProduct(_, resp) {
@@ -12,12 +13,17 @@ class ProductController {
   }
 
   async getProduct(req, resp) {
+    const isIdValid = ObjectID.isValid(req.params.id);
     try {
-      const product = await Product.findById(req.params.id);
-      if (product !== null) {
-        resp.status(200).json(product);
+      if (isIdValid) {
+        const product = await Product.findById(req.params.id);
+        if (product !== null) {
+          resp.status(200).json(product);
+        } else {
+          resp.status(404).json({ message: "Product not found" });
+        }
       } else {
-        resp.status(404).json({ message: "Product not found" });
+        return resp.status(422).json({ message: "ObjectId is not valid" });
       }
     } catch (error) {
       resp.status(500).json({ message: error.message });
@@ -42,33 +48,43 @@ class ProductController {
   }
 
   async updateProduct(req, resp) {
+    const isIdValid = ObjectID.isValid(req.params.id);
     try {
-      const checkProduct = await Product.findById(req.params.id);
-      if (checkProduct !== null) {
-        const product = await Product.find({ model: req.body.model });
-        const isExist = !_.isEmpty(product) ? true : false;
-        if (!isExist) {
-          await Product.findByIdAndUpdate(req.params.id, req.body);
-          resp.status(200).json(await Product.findById(req.params.id));
+      if (isIdValid) {
+        const checkProduct = await Product.findById(req.params.id);
+        if (checkProduct !== null) {
+          const product = await Product.find({ model: req.body.model });
+          const isExist = !_.isEmpty(product) ? true : false;
+          if (!isExist) {
+            await Product.findByIdAndUpdate(req.params.id, req.body);
+            resp.status(200).json(await Product.findById(req.params.id));
+          } else {
+            return resp.status(400).json({ message: "Product already exist" });
+          }
         } else {
-          return resp.status(400).json({ message: "Product already exist" });
+          return resp.status(404).json({ message: "Product not found" });
         }
       } else {
-        return resp.status(404).json({ message: "Product not found" });
+        return resp.status(422).json({ message: "ObjectId is not valid" });
       }
     } catch (error) {
-      return resp.status(500).json({ message: error.message });
+      resp.status(500).json({ message: error.message });
     }
   }
 
   async deleteProduct(req, resp) {
+    const isIdValid = ObjectID.isValid(req.params.id);
     try {
-      const product = await Product.findById(req.params.id);
-      if (product === null) {
-        return resp.status(404).json({ message: "Product not found" });
+      if (isIdValid) {
+        const product = await Product.findById(req.params.id);
+        if (product === null) {
+          return resp.status(404).json({ message: "Product not found" });
+        }
+        await Product.findByIdAndDelete(req.params.id);
+        resp.status(200).json({ message: "Product deleted successfully" });
+      } else {
+        return resp.status(422).json({ message: "ObjectId is not valid" });
       }
-      await Product.findByIdAndDelete(req.params.id);
-      resp.status(200).json({ message: "Product deleted successfully" });
     } catch (error) {
       resp.status(500).json({ message: error.message });
     }
